requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asgiref==3.5.2
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
autopage==0.5.1
backports.cached-property==1.0.2
backports.zoneinfo;python_version<"3.9"
bert-extractive-summarizer==0.10.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cliff==4.1.0
cmaes==0.9.0
cmd2==2.4.2
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
decorator==5.1.1
Django==4.1.4
django-cors-headers==4.0.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.4
einops==0.8.0
ffmpeg-python==0.2.0
filelock==3.8.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2024.3.1
future==0.18.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.51.1
hmmlearn==0.2.8
HyperPyYAML==1.1.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mpmath==1.2.1
multidict==6.0.3
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
optuna==3.6.1
packaging==21.3
pandas==1.5.2
pbr==5.11.0
Pillow==9.3.0
pooch==1.6.0
prettytable==3.5.0
primePy==1.3
protobuf==3.20.1
pyannote.audio==3.0.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-speech-features==0.6
pytorch-lightning==2.2.4
pytorch-metric-learning==2.5.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.8.1
semver==3.0.2
sentence-transformers==2.7.0
sentencepiece==0.1.97
shellingham==1.5.0
simplejson==3.18.0
singledispatchmethod==1.0
six==1.16.0
sortedcontainers==2.4.0
SoundFile==0.12.1
speechbrain==1.0.0
SpeechRecognition==3.9.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
stevedore==4.1.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==2.3.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.3.0
torchmetrics==1.4.0
torchvision==0.18.0
tqdm==4.64.1
transformers==4.40.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
